Adiciona controlador REST para todas as classes

+++ src/main/java/rentvent/ClientesRepo.java
@@ -2,6 +2,6 @@ package rentvent;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClientesRepo extends CrudRepository<Clientes, Integer> {
+public interface ClientesRepo extends CrudRepository<Clientes, Long> {
     
 }
